adding preprocessor for debug and fixed a warning in variant.h about conversion from size_t to int.

+++ src/variant.h
@@ -179,7 +179,7 @@ namespace fcl
 		}
 
 		void * ptr_;
-		size_t index_;
+		int index_;
 	};
 
 	//variant variant traits implenmentation
